:zap: improve folder type and add router
