+++ src/App.js
@@ -1,25 +1,28 @@
 import { Reset } from "styled-reset";
-import styled, { createGlobalStyle } from "styled-components";
-import Header from "./Header"
+import GlobalStyles from "./style/GlobalStyle";
+import Header from "./Components/Header/Header";
 import QuestionPage from "./QuestionPage";
-
-const GlobalStyles = createGlobalStyle`
-  @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@300,400;700&display=swap');
-  body{
-    background: #2d2d2d;
-    color:#fff;
-    font-family: 'Roboto', sans-serif;
-  }
-`;
-
+import { Route, Switch, Redirect } from "react-router-dom";
+import LoginPage from "./Screens/LoginPage";
+import HomePage from "./Screens/HomePage";
 function App() {
+  const routes = (
+    <Switch>
+      <Route exact path="/HomePage" component={HomePage} />
+      <Route exact path="/LoginPage" component={LoginPage} />
+      <Redirect to="/" />
+    </Switch>
+  );
+
   return (
-   <div>
-     <Reset />
-     <GlobalStyles/>
-     <Header/>
-     <QuestionPage />
-   </div>
+    <div>
+      <GlobalStyles />
+      <Header />
+      <div>{routes}</div>
+
+      {/* <Reset /> */}
+      {/* <QuestionPage /> */}
+    </div>
   );
 }
