+++ src/Components/Input/Inputs.js
@@ -1,0 +1,23 @@
+import styled from "styled-components";
+
+export const Input = styled.input`
+  display: inline-block;
+  box-sizing: border-box;
+  width: 100%;
+  border-radius: 3px;
+  border: 1px solid #777;
+  background: rgba(0, 0, 0, 0.1);
+  padding: 8px 10px;
+  margin-top: 9px;
+`;
+
+export const SearchInput = styled.input`
+  display: inline-block;
+  box-sizing: border-box;
+  width: 100%;
+  border-radius: 3px;
+  border: 1px solid #777;
+  background: rgba(0, 0, 0, 0.1);
+  padding: 8px 10px;
+  margin-top: 9px;
+`;
