+++ src/Screens/HomePage.js
@@ -1,0 +1,11 @@
+import Header from "../Components/Header/Header";
+
+function HomePage() {
+  return (
+    <div>
+      <div>HomePage</div>
+    </div>
+  );
+}
+
+export default HomePage;
