+++ src/Screens/LoginPage.js
@@ -1,0 +1,11 @@
+import Header from "../Components/Header/Header";
+
+function LoginPage() {
+  return (
+    <div>
+      <div>LoginPage</div>
+    </div>
+  );
+}
+
+export default LoginPage;
