+++ src/style/GlobalStyle.js
@@ -1,0 +1,11 @@
+import { createGlobalStyle } from "styled-components";
+const GlobalStyles = createGlobalStyle`
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@300,400;700&display=swap');
+body{
+  background: #2d2d2d;
+  color:#fff;
+  font-family: 'Roboto', sans-serif;
+}
+`;
+
+export default GlobalStyles;
